Refactoring: move RegistrationFacade and his cascade dependency to framework bundle
