+++ packages/framework/src/Model/Customer/User/RegistrationDataFactory.php
@@ -1,0 +1,41 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shopsys\FrameworkBundle\Model\Customer\User;
+
+use Shopsys\FrameworkBundle\Component\Domain\Domain;
+use Shopsys\FrameworkBundle\Model\Country\CountryFacade;
+
+class RegistrationDataFactory
+{
+    /**
+     * @param \Shopsys\FrameworkBundle\Component\Domain\Domain $domain
+     * @param \Shopsys\FrameworkBundle\Model\Country\CountryFacade $countryFacade
+     */
+    public function __construct(
+        protected readonly Domain $domain,
+        protected readonly CountryFacade $countryFacade,
+    ) {
+    }
+
+    /**
+     * @param int $domainId
+     * @return \Shopsys\FrameworkBundle\Model\Customer\User\RegistrationData
+     */
+    public function createForDomainId(int $domainId): RegistrationData
+    {
+        $registrationData = $this->create();
+        $registrationData->domainId = $domainId;
+
+        return $registrationData;
+    }
+
+    /**
+     * @return \Shopsys\FrameworkBundle\Model\Customer\User\RegistrationData
+     */
+    public function create(): RegistrationData
+    {
+        return new RegistrationData();
+    }
+}
