+++ packages/frontend-api/src/Model/Order/Exception/OrderCannotBePairedException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\FrontendApi\Model\Order\Exception;
+namespace Shopsys\FrontendApiBundle\Model\Order\Exception;
 
 use Overblog\GraphQLBundle\Error\UserError;
 use Shopsys\FrameworkBundle\Model\Order\Exception\OrderException;
@@ -10,13 +10,13 @@ use Shopsys\FrontendApiBundle\Model\Error\UserErrorWithCodeInterface;
 
 class OrderCannotBePairedException extends UserError implements OrderException, UserErrorWithCodeInterface
 {
-    private const CODE = 'order-cannot-be-paired-with-new-registration';
+    protected const CODE = 'order-cannot-be-paired-with-new-registration';
 
     /**
      * {@inheritdoc}
      */
     public function getUserErrorCode(): string
     {
-        return self::CODE;
+        return static::CODE;
     }
 }
